+++ deploy/sdk/src/dynamo/sdk/cli/utils.py
@@ -28,14 +28,15 @@ import socket
 from typing import Any, DefaultDict, Dict, Iterator, Optional, Protocol, TextIO, Union
 
 import click
+import typer
 import yaml
 from click import Command, Context
 from rich.console import Console
 
-from components.planner import PlannerDefaults  # type: ignore[attr-defined]
+from dynamo.planner.defaults import PlannerDefaults  # type: ignore[attr-defined]
 from dynamo.runtime.logging import configure_dynamo_logging
+from dynamo.sdk.core.protocol.interface import ComponentType
 from dynamo.sdk.core.runner import TargetEnum
-from dynamo.sdk.lib.service import ComponentType
 
 configure_dynamo_logging()
 
@@ -387,7 +388,7 @@ def is_local_planner_enabled(svc: Any, service_configs: dict) -> bool:
         bool: True if local planner is enabled, False otherwise
     """
     # Check all nodes to find planner
-    nodes = [dep.on for dep in svc.dependencies.values()]
+    nodes = [dep for dep in svc.all_services().values()]
     nodes.append(svc)
     planners = [
         node
@@ -399,7 +400,7 @@ def is_local_planner_enabled(svc: Any, service_configs: dict) -> bool:
         console.print(
             "[bold red]Error:[/bold red] More than one planner found in the pipeline"
         )
-        raise click.Exit(code=1)
+        raise typer.Exit(code=1)
 
     # Exactly one planner
     if planners:
@@ -424,7 +425,7 @@ def raise_local_planner_warning(svc: Any, service_configs: dict) -> None:
     no_op = planner_config.get("no-operation", PlannerDefaults.no_operation)
 
     # Check worker counts across nodes
-    nodes = [dep.on for dep in svc.dependencies.values()]
+    nodes = [dep for dep in svc.all_services().values()]
     nodes.append(svc)
     worker_names = ("PrefillWorker", "VllmWorker")
     worker_counts_greater_than_one = [
@@ -432,6 +433,7 @@ def raise_local_planner_warning(svc: Any, service_configs: dict) -> None:
     ]
 
     if any(worker_counts_greater_than_one) and not no_op:
-        logger.warning(
+        logger.error(
             "Local planner is enabled, but workers for prefill or decode is > 1. Local planner must be started with prefill and decode workers set to 1."
         )
+        raise typer.Exit(code=1)
