+++ deploy/sdk/src/dynamo/sdk/core/runner/dynamo.py
@@ -19,6 +19,7 @@ import logging
 import os
 import shlex
 import sys
+from dataclasses import asdict
 from typing import Any, Dict, List, Optional, Set, Type, TypeVar
 
 import psutil
@@ -80,6 +81,8 @@ class LocalService(ServiceMixin, ServiceInterface[T]):
         self._dynamo_config = dynamo_config or DynamoConfig(
             name=name, namespace="default"
         )
+        # Add the dynamo config to the service config
+        self._config["dynamo"] = asdict(self._dynamo_config)
         self._watcher = watcher
         self._socket = socket
         self.app = app or FastAPI(title=name)
